Address review: move ExampleSecondaryTier from __init__.py to manager.py
